// src/app/admin/data-access/models/details-model.ts
export interface Details {
  year?: any;
  universities?: any;
  id?: number; // change to university id
  fullName: string;
  date: string;
  amount: number;
  applicationId: number;
  status: string;
  applicationGuid: string;
  accommodation?: number;
  meals?: number;
  other?: number;
  tuition?: number;
  documentsId?: number;
  bursaryType?: string;
  universityName?:string;
}